refactor: a lot:   - rewrite of creates_sprite   – reverted to src layout (pip prefers)
