+++ PyScratch/src/PyScratch/variable.py
@@ -1,14 +1,12 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from PyScratch.PyScratch import AutoId
-
-from PyScratch.PyScratch import PrimitiveBlock
+from .primitives import PrimitiveBlock
+from .utils import AutoId
 
 
 @dataclass
 class Variable(AutoId):
-
     name: str
     value: str
     is_cloud: bool
