+++ ScratchScript/function_specs/hashes.json
