+++ ScratchScript/src/ScratchScript/lang_compiler/README.md
@@ -15,8 +15,13 @@
       - types for all the lang_extract.py stuff
       - extracted form lang_extract.py because sprites can be arguments
 
+lang_code - parse code
+lang_functions - ABC for functions
+lang_extract - main
+lang_types - types
+
 ## TODO:
- - [ ] rewrite lang_code.parse_fn_call
+ - [x] rewrite lang_code.parse_fn_call
  - [ ] make lang_code.parse_expr
  - [ ] rethink parsing fn calls
- - [ ] does we need specs in lang_functions?
+ - [x] does we need specs in lang_functions? No
