+++ ScratchScript/src/ScratchScript/lang_compiler/__init__.py
@@ -1,0 +1,6 @@
+from .lang_extract import parse_program
+
+
+def main(program):
+    program = parse_program(program)
+
